+++ CHANGELOG.md
@@ -14,6 +14,11 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Add `ua::X509IdentityToken` and `ua::IssuedIdentityToken`.
 - Breaking: Add enum variants `X509` and `Issued` to `UserIdentityToken`.
 
+### Changed
+
+- Breaking: Expect `ua::String`, `ua::ByteString` in `ua::UserNameIdentityToken::with_user_name()`,
+  `ua::UserNameIdentityToken::with_password()`.
+
 ## [0.7.3]
 
 ### Added